refactor(startup): split the restore loop; finish the pre-existing cleanup
Checkpoint 4, closing the last two findings.

#6  `restoreWorkspaceTabs` was 83 lines running five interleaved policies at one
    nesting level: media routing, read failure, stale dedup, create, ingest
    failure. Each has its own reason and they shared one `try`, which is how one
    of them -- the stale dedup -- stayed a live defect while looking ordinary.
    Now a 43-line loop over `restoreOnePath`, which returns whether it created.

#7  `WindowContext`'s init: the storage-preparation step moved to
    `prepareWindowStorage`. Its internal ORDER is load-bearing (the storage key
    must be chosen before `rehydrate()`, or the store loads the previous
    window's workspace), which is exactly why it is one named function rather
    than four statements buried mid-init.

WindowContext.tsx is 348 -> 299 lines and is now PRUNED from the file-size
baseline entirely: it is under the project's 300-line limit, so it is no longer
a violator to be capped. No baseline entry was ever raised in this pass; two
were lowered and one removed.

Verified: `pnpm check:all` exit 0 -- check:static, test:coverage, check:servers,
check:build. 35,653 app tests, 966 gate tests, 512 sidecar, 159 content-server.

The `❌ Trusted-HTML contract broken` lines in that log are the trusted-html
gate's OWN self-test spawning itself against deliberately broken fixtures and
asserting those messages; the real run reports `✅ Trusted-HTML contract OK`.
Checked rather than assumed, because a red-looking line in a green run is
exactly the thing worth explaining.
